requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
firebase-admin==6.2.0
werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
email-validator==2.1.0.post1
Flask-Mail==0.9.1
Authlib==1.3.1
requests==2.31.0
